# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  model.AdminDashboardData:
    properties:
      active_users:
        type: integer
      monitored_sites:
        type: integer
      recent_errors:
        items:
          $ref: '#/definitions/model.ScrapingError'
        type: array
      scraping_errors:
        type: integer
      total_revenue:
        type: number
    type: object
  model.AnalyzeJobRequest:
    properties:
      job_id:
        example: 42
        type: integer
    required:
    - job_id
    type: object
...
  model.Curriculum:
    properties:
      educations:
        items:
          $ref: '#/definitions/model.Education'
        type: array
      experiences:
        items:
          $ref: '#/definitions/model.Experience'
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      languages:
        type: string
      skills:
        type: string
      summary:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
  model.DashboardData:
    properties:
      alerts_sent_count:
        type: integer
      latest_jobs:
        items:
          $ref: '#/definitions/model.Job'
        type: array
      monitored_urls_count:
        type: integer
      new_jobs_today_count:
        type: integer
      user_monitored_urls:
        items:
          $ref: '#/definitions/model.MonitoredURL'
        type: array
    type: object
  model.Education:
    properties:
      degree:
        type: string
      institution:
        type: string
      year:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        example: mensagem de erro
        type: string
    type: object
  model.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
  model.Gap:
    properties:
      areaDescription:
        type: string
      jobRequirementImpacted:
        type: string
    type: object
...
  model.Job:
    properties:
      company:
        type: string
      description:
        type: string
      id:
        type: integer
      job_id:
        type: integer
      job_link:
        type: string
      location:
        type: string
      site_id:
        type: integer
      title:
        type: string
    type: object
...
  model.MatchAnalysis:
    properties:
      overallScoreNumeric:
        type: integer
      overallScoreQualitative:
        type: string
      summary:
        type: string
    type: object
  model.MessageResponse:
    properties:
      message:
        example: operação realizada com sucesso
        type: string
    type: object
  model.MonitoredURL:
    properties:
      base_url:
        type: string
      site_name:
        type: string
    type: object
...
  model.PaginatedJobs:
    properties:
      jobs:
        items:
          $ref: '#/definitions/model.Job'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
    type: object
...
  model.ResumeAnalysis:
    properties:
      actionableResumeSuggestions:
        items:
          $ref: '#/definitions/model.Suggestion'
        type: array
      finalConsiderations:
        type: string
      gapsAndImprovementAreas:
        items:
          $ref: '#/definitions/model.Gap'
        type: array
      matchAnalysis:
        $ref: '#/definitions/model.MatchAnalysis'
      strengthsForThisJob:
        items:
          $ref: '#/definitions/model.Strength'
        type: array
    type: object
...
  model.ScrapingError:
    properties:
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: integer
      site_name:
        type: string
    type: object
  model.SendAnalysisEmailRequest:
    properties:
      analysis:
        $ref: '#/definitions/model.ResumeAnalysis'
      job_id:
        example: 42
        type: integer
    required:
    - analysis
    - job_id
    type: object
...
  model.Strength:
    properties:
      point:
        type: string
      relevanceToJob:
        type: string
    type: object
  model.Suggestion:
    properties:
      curriculumSectionToApply:
        type: string
      exampleWording:
        type: string
      reasoningForThisJob:
        type: string
      suggestion:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para plataforma de scraping e matching de vagas de emprego.
  title: ScrapJobs API
  version: "1.0"
paths:
  /api/admin/dashboard:
    get:
      description: Retorna metricas administrativas (somente admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AdminDashboardData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Dashboard administrativo
      tags:
      - Admin
  /api/analyze-job:
    post:
      consumes:
      - application/json
      description: Analisa compatibilidade do curriculo com uma vaga usando IA
      parameters:
      - description: ID da vaga
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AnalyzeJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResumeAnalysis'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Analisar vaga com IA
      tags:
      - Analysis
  /api/analyze-job/send-email:
    post:
      consumes:
      - application/json
      description: Envia resultado da analise de compatibilidade por email
      parameters:
      - description: ID da vaga e resultado da analise
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SendAnalysisEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Enviar analise por email
      tags:
      - Analysis
  /api/dashboard:
    get:
      description: Retorna estatisticas e vagas recentes do usuario
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DashboardData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Dados do dashboard
      tags:
      - Dashboard
  /api/dashboard/jobs:
    get:
      description: Retorna vagas com paginacao e filtros
      parameters:
      - default: 1
        description: Pagina
        in: query
        name: page
        type: integer
      - default: 10
        description: Limite por pagina (max 50)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Filtrar por dias
        in: query
        name: days
        type: integer
      - description: Buscar por titulo
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedJobs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Vagas recentes paginadas
      tags:
      - Dashboard
...
  /curriculum:
    get:
      description: Retorna todos os curriculos do usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Curriculum'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Listar curriculos
      tags:
      - Curriculum
    post:
      consumes:
      - application/json
      description: Cria um novo curriculo para o usuario autenticado
      parameters:
      - description: Dados do curriculo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Curriculum'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Curriculum'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Criar curriculo
      tags:
      - Curriculum
  /curriculum/{id}:
    put:
      consumes:
      - application/json
      description: Atualiza um curriculo existente
      parameters:
      - description: ID do curriculo
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Curriculum'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Curriculum'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Atualizar curriculo
      tags:
      - Curriculum
  /curriculum/{id}/active:
    patch:
      description: Define um curriculo como ativo para o usuario
      parameters:
      - description: ID do curriculo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Ativar curriculo
      tags:
      - Curriculum
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: Authorization
    type: apiKey
swagger: "2.0"